planetarium.html
Attn Shara-

I based the new planetarium.html on the registration page as you suggested. OLD page was renamed so you have a backup of the original layout.

planetarium.html has the form, with extra stuff removed (the previous form, the extra image).

I did NOT alter the form to do the correct actions. I am unsure how to do that.

The form and container you see in the middle needs to move right so the copy can be moved up to inhabit the left side of the page.

I will make a new branch after you merge this.  I will research those things I can't do yet, but I wanted to get this to you for now, in case you already know how to manipulate these objects.